+++ Arbiter.App/ViewModels/Proxy/ProxyViewModel.Debug.cs
@@ -1,48 +1,27 @@
 ﻿using Arbiter.App.Models;
-using Microsoft.Extensions.Logging;
+using Arbiter.Net.Client.Messages;
+using Arbiter.Net.Server.Messages;
 
 namespace Arbiter.App.ViewModels.Proxy;
 
 public partial class ProxyViewModel
 {
+    private readonly IClientMessageFactory _clientMessageFactory = new ClientMessageFactory();
+    private readonly IServerMessageFactory _serverMessageFactory = new ServerMessageFactory();
+
     public void ApplyDebugFilters(DebugSettings settings)
     {
         RemoveDebugFilters();
 
-        if (settings.ShowNpcId || settings.ShowMonsterId || settings.ShowMonsterClickId)
-        {
-            AddDebugEntityFilters(settings);
-        }
+        // Ideally, these would all be moved to Lua scripts
+        // For now they are done in code for simplicity
 
-        if (settings.ShowDialogId)
-        {
-            AddDebugDialogFilters(settings);
-        }
-
-        if (settings.ShowHiddenPlayers || settings.ShowPlayerNames)
-        {
-            AddDebugPlayerFilters(settings);
-        }
-
-        if (settings.DisableBlind)
-        {
-            AddDebugEffectsFilters(settings);
-        }
-
-        if (settings.EnableTabMap || settings.DisableWeatherEffects || settings.DisableDarkness)
-        {
-            AddDebugMapFilters(settings);
-        }
-
-        if (settings.IgnoreEmptyMessages)
-        {
-            AddDebugMessageFilters(settings);
-        }
-
-        if (settings.CheckEnabled())
-        {
-            _logger.LogInformation("Debug packet filters enabled");
-        }
+        AddDebugEntityFilters(settings);
+        AddDebugDialogFilters(settings);
+        AddDebugPlayerFilters(settings);
+        AddDebugEffectsFilters(settings);
+        AddDebugMapFilters(settings);
+        AddDebugMessageFilters(settings);
     }
 
     public void RemoveDebugFilters()
@@ -53,7 +32,5 @@ public partial class ProxyViewModel
         RemoveDebugEffectsFilters();
         RemoveDebugMapFilters();
         RemoveDebugMessageFilters();
-
-        _logger.LogInformation("Debug packet filters disabled");
     }
 }